+++ util-networks-metrics.R
@@ -16,6 +16,7 @@
 ## Copyright 2017-2019 by Claus Hunsen <hunsen@fim.uni-passau.de>
 ## Copyright 2017-2018 by Christian Hechtl <hechtl@fim.uni-passau.de>
 ## Copyright 2018 by Barbara Eckl <ecklbarb@fim.uni-passau.de>
+## Copyright 2021 by Niklas Schneider <s8nlschn@stud.uni-saarland.de>
 ## All Rights Reserved.
 
 
@@ -31,10 +32,19 @@ requireNamespace("igraph")
 #' Determine the maximum degree for the given network.
 #'
 #' @param network the network to be examined
-#' @param mode the mode to be used for determining the degrees
+#' @param mode the mode to be used for determining the degrees [default: "total"]
 #'
-#' @return A dataframe containing the name of the vertex with with maximum degree its degree.
+#' @return A data frame containing the name of the vertex with with maximum degree its degree.
 metrics.hub.degree = function(network, mode = c("total", "in", "out")) {
+    ## check whether the network is empty, i.e., if it has no vertices
+    if (igraph::vcount(network) == 0) {
+        ## print user warning instead of igraph error
+        logging::logwarn("The input network has no vertices. Will return NA right away.")
+
+        ## cancel the execution and return NA
+        return(NA)
+    }
+
     mode = match.arg(mode)
     degrees = igraph::degree(network, mode = c(mode))
     vertex = which.max(degrees)
@@ -45,7 +55,7 @@ metrics.hub.degree = function(network, mode = c("total", "in", "out")) {
 #' Calculate the average degree of a network.
 #'
 #' @param network the network to be examined
-#' @param mode the mode to be used for determining the degrees
+#' @param mode the mode to be used for determining the degrees [default: "total"]
 #'
 #' @return The average degree of the vertices in the network.
 metrics.avg.degree = function(network, mode = c("total", "in", "out")) {
@@ -58,9 +68,9 @@ metrics.avg.degree = function(network, mode = c("total", "in", "out")) {
 #' Calculate all vertex degrees for the given network
 #'
 #' @param network the network to be examined
-#' @param sort whether the resulting dataframe is to be sorted by the vertex degree
+#' @param sort whether the resulting dataframe is to be sorted by the vertex degree [default: TRUE]
 #' @param sort.decreasing if sorting is active, this says whether the dataframe is to be
-#'            sorted in descending or ascending order
+#'                        sorted in descending or ascending order [default: TRUE]
 #'
 #' @return A dataframe containing the vertices and their respective degrees.
 metrics.vertex.degrees = function(network, sort = TRUE, sort.decreasing = TRUE) {
@@ -85,11 +95,14 @@ metrics.density = function(network) {
 #' Calculate the average path length for the given network.
 #'
 #' @param network the network to be examined
-#' @param directed whether to consider directed paths in directed networks
-#' @param unconnected whether all vertices of the network are connected
+#' @param directed whether to consider directed paths in directed networks [default: TRUE]
+#' @param unconnected whether there are subnetworks in the network that are not connected.
+#'                    If \code{TRUE} only the lengths of the existing paths are considered and averaged;
+#'                    if \code{FALSE} the length of the missing paths are counted having length \code{vcount(graph)}, one longer than
+#'                    the longest possible geodesic in the network (from igraph documentation) [default: TRUE]
 #'
-#' @return The average pathlength of the given network.
-metrics.avg.pathlength = function(network, directed, unconnected) {
+#' @return The average path length of the given network.
+metrics.avg.pathlength = function(network, directed = TRUE, unconnected = TRUE) {
     avg.pathlength = igraph::average.path.length(network, directed = directed, unconnected = unconnected)
     return(c(avg.pathlength = avg.pathlength))
 }
@@ -100,7 +113,7 @@ metrics.avg.pathlength = function(network, directed, unconnected) {
 #' Such vertices are removed from all average calculations for any averaging \code{cc.type}.
 #'
 #' @param network the network to be examined
-#' @param cc.type the type of cluserting coefficient to be calculated
+#' @param cc.type the type of cluserting coefficient to be calculated [default: "global"]
 #'
 #' @return The clustering coefficient of the network.
 metrics.clustering.coeff = function(network, cc.type = c("global", "local", "barrat", "localaverage")) {
@@ -113,7 +126,7 @@ metrics.clustering.coeff = function(network, cc.type = c("global", "local", "bar
 #'
 #' @param network the network to be examined
 #' @param community.detection.algorithm the algorithm to be used for the detection of communities
-#'            which is required for the calculation of the clustering coefficient
+#'            which is required for the calculation of the clustering coefficient [default: igraph::cluster_walktrap]
 #'
 #' @return The modularity value for the given network.
 metrics.modularity = function(network, community.detection.algorithm = igraph::cluster_walktrap) {
@@ -133,8 +146,8 @@ metrics.modularity = function(network, community.detection.algorithm = igraph::c
 #' The algorithm relies on the Erdös-Renyi random network with the same number
 #' of vertices and edges as the given network.
 #'
-#' To check the result value \code{s.delta} for a binary (true/false) decision on smallworldness,
-#' do this: \code{is.smallworld = s.delta > 1}.
+#' In order to get a binary (true/false) decision on smallworldness of a network,
+#' use \code{metrics.is.smallworld} instead.
 #'
 #' Important: The given network needs to be simplified for the calculation to work!
 #'
@@ -142,31 +155,53 @@ metrics.modularity = function(network, community.detection.algorithm = igraph::c
 #'
 #' @return The smallworldness value of the network.
 metrics.smallworldness = function(network) {
-    # construct Erdös-Renyi network with same number of vertices and edges as g
+    ## first check whether the network is simplified
+    if (!is.simple(network)) {
+        ## if this is not the case, raise an error and stop the execution
+        error.message = "The input network has too many edges. Try again with a simplified network."
+        logging::error(error.message)
+        stop(error.message)
+    }
+
+    ## else construct Erdös-Renyi network 'h' with same number of vertices and edges as the given network 'network',
+    ## as the requirement of the function is fulfilled
     h = igraph::erdos.renyi.game(n = igraph::vcount(network),
                                  p.or.m = igraph::ecount(network),
                                  type = "gnm",
                                  directed = FALSE)
 
-    # compute clustering coefficients
+    ## compute clustering coefficients
     g.cc = igraph::transitivity(network, type = "global")
     h.cc = igraph::transitivity(h, type = "global")
-    # compute average shortest-path length
+    ## compute average shortest-path length
     g.l = igraph::average.path.length(network, unconnected = TRUE)
     h.l = igraph::average.path.length(h, unconnected = TRUE)
 
-    # binary decision
-    # intermediate variables
+    ## binary decision
+    ## intermediate variables
     gamma = g.cc / h.cc
     lambda = g.l / h.l
 
-    # indicator s.delta
+    ## indicator s.delta
     s.delta = gamma / lambda
 
-    ## if s.delta > 1, then the network is a small-world network
-    # is.smallworld = s.delta > 1
     return (c(smallworldness = s.delta))
 }
+
+#' Decide, whether a network is smallworld or not.
+#'
+#' @param network the network to be examined
+#'
+#' @return \code{TRUE}, if the network is smallworld,
+#'         \code{FALSE}, if it is not,
+#'         \code{NA}, if an error occured.
+metrics.is.smallworld = function(network) {
+    s.delta = metrics.smallworldness(network)
+
+    ## return whether the network is smallworld
+    return(s.delta > 1)
+}
+
 
 #' Determine scale freeness of a network using the power law fitting method.
 #'
@@ -233,7 +268,7 @@ metrics.is.scale.free = function(network, minimum.number.vertices = 30) {
 #'
 #' @param network the network to be examined
 #'
-#' @return A data.frame containing the following columns:
+#' @return a data.frame containing the following columns:
 #'         - \code{"deg"}: the vertex degrees for all vertices in the given network,
 #'         - \code{"cc"}: the local clustering coefficient for all vertices in the given network, and
 #'         - \code{"log.deg"} and \code{"log.cc"}: the logarithmic values for the columns
@@ -250,3 +285,88 @@ metrics.hierarchy = function(network) {
 }
 
 
+## The column headers for a centrality data frame calculated by the function \code{metrics.vertex.centralities}
+VERTEX.CENTRALITIES.COLUMN.NAMES = c("vertex.name", "centrality")
+
+#' Calculate the centrality value for vertices from a network and project data.
+#' If a \code{ProjectData} is supplied, only vertices from the network that are also present in the project data are
+#' considered. Otherwise, if no custom vector \code{restrict.classification.to.vertices} is supplied, all vertices of
+#' the network are considered.
+#'
+#' @param network the network containing the vertices to classify
+#' @param proj.data the \code{ProjectData} containing the authors or artifacts to classify
+#' @param type a character string declaring the classification metric. The classification metric determines which
+#'             numerical characteristic of vertices is chosen as their centrality value.
+#'             The parameter only supports network-based options/metrics:
+#'              - "network.degree"
+#'              - "network.eigen"
+#'              - "network.hierarchy"
+#'             [defalt: "network.degree"]
+#' @param restrict.classification.to.vertices a vector of vertex names. Only vertices that are contained within this
+#'                                            vector are to be classified. Vertices that appear in the vector but are
+#'                                            not part of the classification result (i.e., they are not present in the
+#'                                            underlying data) will be added to it afterwards (with a centrality value
+#'                                            of \code{NA}). \code{NULL} means that the restriction is automatically
+#'                                            calculated from the data based on the network's edge relations if and only
+#'                                            if both network and data are present. In any other case \code{NULL} will
+#'                                            not introduce any further restriction. [default: NULL]
+#' @return a data.frame with the columns \code{"vertex.name"} and \code{"centrality"} containing the centrality values
+#'         for each respective vertex
+metrics.vertex.centralities = function(network,
+                              proj.data,
+                              type = c("network.degree",
+                                       "network.eigen",
+                                       "network.hierarchy"),
+                              restrict.classification.to.vertices = NULL) {
+    type = match.arg(type)
+
+    ## check whether the restrict parameter is set to default 'NULL'
+    if (is.null(restrict.classification.to.vertices)) {
+        ## now check whether both data and network are present
+        if (!is.null(network) && !is.null(proj.data)) {
+            ## in this case calculate the restrict parameter based on the edge relation along with the vertices from
+            ## these data.sources
+
+            ## first check whether the network only consists of author vertices
+            ## therefore get a vector with all vertex types
+            vertex.types = unique(igraph::V(network)$type)
+            if (vertex.types == c("author")) {
+                ## in this case, use the 'get.authors.by.data.source' function to get the author list
+                restrict.classification.to.vertices = get.authors.by.data.source(
+                    get.data.sources.from.relations(network),
+                    proj.Data)
+            }
+            else if (vertex.types == c("artifact")) {
+                ## in this case, always use artifact relation, as both unipartite edges connecting to artifact vertices
+                ## as well as bipartite have an artifact relation
+                restrict.classification.to.vertices = get.artifacts(get.data.sources.from.relations(network))
+            }
+            else {
+                ## when both vertex types are present, compute both authors and artifacts into one vector
+                restrict.authors = get.authors.by.data.source(get.data.sources.from.relations(network),
+                                                              proj.Data)
+                restrict.artifacts = get.artifacts(get.data.sources.from.relations(network))
+                restrict.classification.to.vertices = append(restrict.authors, restrict.artifacts)
+            }
+        }
+        ## else leave the parameter at 'NULL' which still serves as a default value for the
+        ## 'get.author.class.by.type' function
+    }
+
+    ## calculate the centrality tables
+    class = get.author.class.by.type(network = network,
+                                     proj.data = proj.data,
+                                     type = type,
+                                     restrict.classification.to.authors = restrict.classification.to.vertices)
+
+    ## bind the two data frames for core and peripheral together
+    centrality = rbind(class[["core"]], class[["peripheral"]])
+
+    ## set column names accordingly
+    colnames(centrality) = VERTEX.CENTRALITIES.COLUMN.NAMES
+
+    ## order by centrality (descending) (with NA being at the bottom) and then by name (ascending)
+    centrality = centrality[order(-centrality[["centrality"]], centrality[["name"]]), ]
+
+    return(centrality)
+}